Update helm version to v2.15.2
Version v2.15.2 includes some incremental fixes.
